Added missing UI
Much needed ui for tables added + made interact durations more precise.

+++ Assets/Scripts/IInteractable.cs
@@ -6,5 +6,5 @@ public interface IInteractable
 {
     float GetXPosition();
     InteractableEnum.Interactables GetInteractable();
-    int GetInteractableDuration();
+    float GetInteractableDuration();
 }
